fix(vite): add dts to build options
