+++ packages/vite/src/generators/configuration/schema.json
@@ -23,8 +23,7 @@
     },
     "includeVitest": {
       "type": "boolean",
-      "description": "Use vitest for the test suite.",
-      "hidden": true
+      "description": "Use vitest for the test suite."
     },
     "uiFramework": {
       "type": "string",
