+++ packages/vite/src/utils/test-files/react-lib-non-buildable-jest.json
