+++ packages/vite/src/utils/vite-config-edit-utils.spec.ts
@@ -1,221 +1,14 @@
-import {
-  readProjectConfiguration,
-  Tree,
-  updateProjectConfiguration,
-} from '@nrwl/devkit';
+import { Tree } from '@nrwl/devkit';
 import { createTreeWithEmptyV1Workspace } from '@nrwl/devkit/testing';
 import { tsquery } from '@phenomnomnominal/tsquery';
-import {
-  ensureBuildOptionsInViteConfig,
-  findExistingTargetsInProject,
-  getViteConfigPathForProject,
-  handleUnsupportedUserProvidedTargets,
-} from './generator-utils';
-import {
-  mockReactAppGenerator,
-  mockViteReactAppGenerator,
-  mockAngularAppGenerator,
-} from './test-utils';
+
+import { ensureBuildOptionsInViteConfig } from './vite-config-edit-utils';
+
 describe('generator utils', () => {
   let tree: Tree;
 
   beforeEach(() => {
     tree = createTreeWithEmptyV1Workspace();
-  });
-
-  describe('getViteConfigPathForProject', () => {
-    beforeEach(() => {
-      mockViteReactAppGenerator(tree);
-    });
-    it('should return correct path for vite.config file if no configFile is set', () => {
-      const viteConfigPath = getViteConfigPathForProject(
-        tree,
-        'my-test-react-vite-app'
-      );
-      expect(viteConfigPath).toEqual(
-        'apps/my-test-react-vite-app/vite.config.ts'
-      );
-    });
-
-    it('should return correct path for vite.config file if custom configFile is set', () => {
-      const projectConfig = readProjectConfiguration(
-        tree,
-        'my-test-react-vite-app'
-      );
-      updateProjectConfiguration(tree, 'my-test-react-vite-app', {
-        ...projectConfig,
-        targets: {
-          ...projectConfig.targets,
-          build: {
-            ...projectConfig.targets.build,
-            options: {
-              ...projectConfig.targets.build.options,
-              configFile: 'apps/my-test-react-vite-app/vite.config.custom.ts',
-            },
-          },
-        },
-      });
-
-      tree.write(`apps/my-test-react-vite-app/vite.config.custom.ts`, '');
-
-      const viteConfigPath = getViteConfigPathForProject(
-        tree,
-        'my-test-react-vite-app'
-      );
-      expect(viteConfigPath).toEqual(
-        'apps/my-test-react-vite-app/vite.config.custom.ts'
-      );
-    });
-
-    it('should return correct path for vite.config file given a target name', () => {
-      const projectConfig = readProjectConfiguration(
-        tree,
-        'my-test-react-vite-app'
-      );
-      updateProjectConfiguration(tree, 'my-test-react-vite-app', {
-        ...projectConfig,
-        targets: {
-          ...projectConfig.targets,
-          'other-build': {
-            ...projectConfig.targets.build,
-            options: {
-              ...projectConfig.targets.build.options,
-              configFile: 'apps/my-test-react-vite-app/vite.other.custom.ts',
-            },
-          },
-        },
-      });
-
-      tree.write(`apps/my-test-react-vite-app/vite.other.custom.ts`, '');
-
-      const viteConfigPath = getViteConfigPathForProject(
-        tree,
-        'my-test-react-vite-app',
-        'other-build'
-      );
-      expect(viteConfigPath).toEqual(
-        'apps/my-test-react-vite-app/vite.other.custom.ts'
-      );
-    });
-  });
-
-  describe('findExistingTargetsInProject', () => {
-    it('should return the correct targets', () => {
-      mockReactAppGenerator(tree);
-      const { targets } = readProjectConfiguration(tree, 'my-test-react-app');
-
-      const existingTargets = findExistingTargetsInProject(targets);
-      expect(existingTargets).toMatchObject({
-        userProvidedTargetIsUnsupported: {
-          build: undefined,
-          serve: undefined,
-          test: undefined,
-        },
-        validFoundTargetName: {
-          build: 'build',
-          serve: 'serve',
-          test: 'test',
-        },
-        projectContainsUnsupportedExecutor: undefined,
-      });
-    });
-
-    it('should return the correct - undefined - targets for Angular apps', () => {
-      mockAngularAppGenerator(tree);
-      const { targets } = readProjectConfiguration(tree, 'my-test-angular-app');
-      const existingTargets = findExistingTargetsInProject(targets);
-      expect(existingTargets).toMatchObject({
-        userProvidedTargetIsUnsupported: {
-          build: undefined,
-          serve: undefined,
-          test: undefined,
-        },
-        validFoundTargetName: {
-          build: undefined,
-          serve: undefined,
-          test: 'test',
-        },
-        projectContainsUnsupportedExecutor: true,
-      });
-    });
-  });
-
-  describe('handleUnsupportedUserProvidedTargets', () => {
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-    it('should throw error if unsupported and user does not want to confirm', async () => {
-      const { Confirm } = require('enquirer');
-      const confirmSpy = jest.spyOn(Confirm.prototype, 'run');
-      confirmSpy.mockResolvedValueOnce(false);
-      expect.assertions(2);
-      const object = {
-        unsupportedUserProvidedTarget: {
-          build: true,
-        },
-        userProvidedTargets: {
-          build: 'my-build',
-          serve: 'my-serve',
-          test: 'my-test',
-        },
-        targets: {
-          build: 'build',
-          serve: 'serve',
-          test: 'test',
-        },
-      };
-
-      try {
-        await handleUnsupportedUserProvidedTargets(
-          object.unsupportedUserProvidedTarget,
-          object.userProvidedTargets,
-          object.targets
-        );
-        throw new Error('should not reach here');
-      } catch (e) {
-        expect(e).toBeDefined();
-        expect(e.toString()).toContain(
-          'The build target my-build cannot be converted to use the @nrwl/vite:build executor'
-        );
-      }
-    });
-
-    it('should NOT throw error if unsupported and user confirms', async () => {
-      const { Confirm } = require('enquirer');
-      const confirmSpy = jest.spyOn(Confirm.prototype, 'run');
-      confirmSpy.mockResolvedValue(true);
-      expect.assertions(2);
-      const object = {
-        unsupportedUserProvidedTarget: {
-          build: true,
-        },
-        userProvidedTargets: {
-          build: 'my-build',
-          serve: 'my-serve',
-          test: 'my-test',
-        },
-        targets: {
-          build: 'build',
-          serve: 'serve',
-          test: 'test',
-        },
-      };
-
-      expect(
-        handleUnsupportedUserProvidedTargets(
-          object.unsupportedUserProvidedTarget,
-          object.userProvidedTargets,
-          object.targets
-        )
-      ).resolves.toBeUndefined();
-
-      const response = await handleUnsupportedUserProvidedTargets(
-        object.unsupportedUserProvidedTarget,
-        object.userProvidedTargets,
-        object.targets
-      );
-      expect(response).toBeUndefined();
-    });
   });
 
   describe('ensureBuildOptionsInViteConfig', () => {
@@ -239,7 +32,6 @@ describe('generator utils', () => {
         external: ["'react', 'react-dom', 'react/jsx-runtime'"]
       }
     },`;
-
     const buildOptionObject = {
       lib: {
         entry: 'src/index.ts',
@@ -251,6 +43,22 @@ describe('generator utils', () => {
         external: ["'react', 'react-dom', 'react/jsx-runtime'"],
       },
     };
+    const dtsPlugin = `dts({
+      tsConfigFilePath: join(__dirname, 'tsconfig.lib.json'),
+      // Faster builds by skipping tests. Set this to false to enable type checking.
+      skipDiagnostics: true,
+    }),`;
+    const dtsImportLine = `import dts from 'vite-plugin-dts';\nimport { join } from 'path';`;
+
+    const pluginOption = `
+    plugins: [
+      ${dtsPlugin}
+      react(),
+      viteTsConfigPaths({
+        root: '../../',
+      }),
+    ],
+    `;
 
     const noBuildOptions = `
     import { defineConfig } from 'vite';
@@ -258,10 +66,6 @@ describe('generator utils', () => {
     import viteTsConfigPaths from 'vite-tsconfig-paths';
 
     export default defineConfig({
-      server: {
-        port: 4200,
-        host: 'localhost',
-      },
       plugins: [
         react(),
         viteTsConfigPaths({
@@ -287,10 +91,6 @@ describe('generator utils', () => {
     import viteTsConfigPaths from 'vite-tsconfig-paths';
 
     export default defineConfig({
-      server: {
-        port: 4200,
-        host: 'localhost',
-      },
       plugins: [
         react(),
         viteTsConfigPaths({
@@ -345,10 +145,6 @@ describe('generator utils', () => {
     import viteTsConfigPaths from 'vite-tsconfig-paths';
 
     export default {
-      server: {
-        port: 4200,
-        host: 'localhost',
-      },
       plugins: [
         react(),
         viteTsConfigPaths({
@@ -368,7 +164,10 @@ describe('generator utils', () => {
         tree,
         'apps/my-app/vite.config.ts',
         buildOption,
-        buildOptionObject
+        buildOptionObject,
+        dtsPlugin,
+        dtsImportLine,
+        pluginOption
       );
       const appFileContent = tree.read('apps/my-app/vite.config.ts', 'utf-8');
       const file = tsquery.ast(appFileContent);
@@ -386,7 +185,10 @@ describe('generator utils', () => {
         tree,
         'apps/my-app/vite.config.ts',
         buildOption,
-        buildOptionObject
+        buildOptionObject,
+        dtsPlugin,
+        dtsImportLine,
+        pluginOption
       );
       const appFileContent = tree.read('apps/my-app/vite.config.ts', 'utf-8');
       const file = tsquery.ast(appFileContent);
@@ -404,7 +206,10 @@ describe('generator utils', () => {
         tree,
         'apps/my-app/vite.config.ts',
         buildOption,
-        buildOptionObject
+        buildOptionObject,
+        dtsPlugin,
+        dtsImportLine,
+        pluginOption
       );
       const appFileContent = tree.read('apps/my-app/vite.config.ts', 'utf-8');
       const file = tsquery.ast(appFileContent);
@@ -422,7 +227,10 @@ describe('generator utils', () => {
         tree,
         'apps/my-app/vite.config.ts',
         buildOption,
-        buildOptionObject
+        buildOptionObject,
+        dtsPlugin,
+        dtsImportLine,
+        pluginOption
       );
       const appFileContent = tree.read('apps/my-app/vite.config.ts', 'utf-8');
       const file = tsquery.ast(appFileContent);
@@ -440,7 +248,10 @@ describe('generator utils', () => {
         tree,
         'apps/my-app/vite.config.ts',
         buildOption,
-        buildOptionObject
+        buildOptionObject,
+        dtsPlugin,
+        dtsImportLine,
+        pluginOption
       );
       const appFileContent = tree.read('apps/my-app/vite.config.ts', 'utf-8');
       const file = tsquery.ast(appFileContent);
@@ -458,7 +269,10 @@ describe('generator utils', () => {
         tree,
         'apps/my-app/vite.config.ts',
         buildOption,
-        buildOptionObject
+        buildOptionObject,
+        dtsPlugin,
+        dtsImportLine,
+        pluginOption
       );
       const appFileContent = tree.read('apps/my-app/vite.config.ts', 'utf-8');
       expect(appFileContent).toMatchSnapshot();
